pyproject.toml
[build-system]
requires = [ "setuptools", "setuptools_scm" ]
build-backend = "setuptools.build_meta"

[project]
name = "vanward"
description = "Scripts to assist prior to certain steps in deploying a new cycle."
license = { text = "GPL3" }
classifiers = [ "Programming Language :: Python :: 3" ]
urls = { documentation = "https://jira.lsstcorp.org/secure/Dashboard.jspa", repository = "https://github.com/lsst-ts/vanward" }
dynamic = [ "version" ]
dependencies = [
  "jira", "gql", "pyyaml", "gitpython"
]

[tool.setuptools.dynamic]
version = { attr = "setuptools_scm.get_version" }

[tool.setuptools.packages.find]
where = [ "python" ]

[project.scripts]
check_conda_package_versions = "lsst.ts.vanward.check_conda_package_versions:runner"
check_software_releases = "lsst.ts.vanward.check_software_releases:runner"
collect_tickets_to_keep = "lsst.ts.vanward.collect_tickets_to_keep:runner"
create_configuration_tickets = "lsst.ts.vanward.create_configuration_tickets:runner"
create_summit_upgrade_ticket = "lsst.ts.vanward.create_summit_upgrade_ticket:runner"
find_merges_without_release_tickets = "lsst.ts.vanward.find_merges_without_release_tickets:runner"
move_bucket_ticket_links = "lsst.ts.vanward.move_bucket_ticket_links:runner"
release_announcement = "lsst.ts.vanward.release_announcement:runner"
release_tickets = "lsst.ts.vanward.release_tickets:runner"

[tool.setuptools_scm]
write_to = "python/lsst/ts/vanward/version.py"
write_to_template = """
# Generated by setuptools_scm
__all__ = ["__version__"]
__version__ = "{version}"
"""

[tool.pytest.ini_options]
addopts = "--black --flake8 --ignore-glob=*/version.py"
flake8-ignore = ["E133", "E203", "E226", "E228", "N802", "N803", "N806", "N812", "N813", "N815", "N816", "W503"]
flake8-max-line-length = 110
flake8-max-doc-length = 79
asyncio_mode = "auto"

[project.optional-dependencies]
dev = [
  "documenteer[pipelines]", "pre-commit"
]
